Align OpenAI okhttp observability in line with Anthropic
  - Remove common utils and move the okhttp observability logic into OpenAI inline with Anthropic implementation
