+++ models/spring-ai-anthropic/src/main/java/org/springframework/ai/anthropic/http/okhttp/SpringAiAnthropicHttpClient.java
@@ -27,7 +27,11 @@ import java.util.Optional;
 import java.util.concurrent.CancellationException;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
+import java.util.concurrent.SynchronousQueue;
+import java.util.concurrent.ThreadFactory;
+import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import javax.net.ssl.HostnameVerifier;
 import javax.net.ssl.SSLSocketFactory;
@@ -44,12 +48,17 @@ import com.anthropic.core.http.HttpRequestBody;
 import com.anthropic.core.http.HttpResponse;
 import com.anthropic.core.http.ProxyAuthenticator;
 import com.anthropic.errors.AnthropicIoException;
+import io.micrometer.context.ContextExecutorService;
+import io.micrometer.context.ContextSnapshotFactory;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.Tag;
+import io.micrometer.core.instrument.binder.okhttp3.OkHttpConnectionPoolMetrics;
+import io.micrometer.core.instrument.binder.okhttp3.OkHttpObservationInterceptor;
 import io.micrometer.observation.ObservationRegistry;
 import okhttp3.Call;
 import okhttp3.Callback;
 import okhttp3.ConnectionPool;
+import okhttp3.Dispatcher;
 import okhttp3.HttpUrl;
 import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
@@ -59,8 +68,6 @@ import okhttp3.Response;
 import okio.BufferedSink;
 import okio.Okio;
 import org.jspecify.annotations.Nullable;
-
-import org.springframework.ai.observation.conventions.SpringAiOkHttpObservabilityUtils;
 
 /**
  * OkHttp-backed {@link HttpClient} for the Anthropic Java SDK, with Micrometer's
@@ -570,8 +577,10 @@ public final class SpringAiAnthropicHttpClient implements HttpClient {
 				.callTimeout(this.timeout.request())
 				.proxy(this.proxy);
 
-			boolean ownsDispatcherExecutor = SpringAiOkHttpObservabilityUtils.configureObservability(okBuilder,
-					this.observationRegistry, OBSERVATION_NAME, this.dispatcherExecutorService);
+			OkHttpObservationInterceptor observationInterceptor = OkHttpObservationInterceptor
+				.builder(this.observationRegistry, OBSERVATION_NAME)
+				.build();
+			okBuilder.addInterceptor(observationInterceptor);
 
 			if (this.proxyAuthenticator != null) {
 				final ProxyAuthenticator pa = this.proxyAuthenticator;
@@ -582,6 +591,14 @@ public final class SpringAiAnthropicHttpClient implements HttpClient {
 					return authed.map(req -> toRequestComputeUrl(req, null)).orElse(null);
 				});
 			}
+
+			ExecutorService userDispatcherExecutor = this.dispatcherExecutorService;
+			boolean ownsDispatcherExecutor = userDispatcherExecutor == null;
+			ExecutorService dispatcherBase = (userDispatcherExecutor != null) ? userDispatcherExecutor
+					: defaultDispatcherExecutor();
+			ExecutorService dispatcherExecutor = ContextExecutorService.wrap(dispatcherBase,
+					ContextSnapshotFactory.builder().build());
+			okBuilder.dispatcher(new Dispatcher(dispatcherExecutor));
 
 			if (this.maxIdleConnections != null && this.keepAliveDuration != null) {
 				okBuilder.connectionPool(new ConnectionPool(this.maxIdleConnections, this.keepAliveDuration.toNanos(),
@@ -609,9 +626,24 @@ public final class SpringAiAnthropicHttpClient implements HttpClient {
 			// the SDK's tuning at the bottom of `OkHttpClient.Builder.build()`.
 			okClient.dispatcher().setMaxRequestsPerHost(okClient.dispatcher().getMaxRequests());
 
-			SpringAiOkHttpObservabilityUtils.bindConnectionPoolMetrics(okClient, this.meterRegistry, this.meterTags);
+			if (this.meterRegistry != null) {
+				new OkHttpConnectionPoolMetrics(okClient.connectionPool(), this.meterTags).bindTo(this.meterRegistry);
+			}
 
 			return new SpringAiAnthropicHttpClient(okClient, resolvedBackend, ownsDispatcherExecutor);
+		}
+
+		// Replicates OkHttp's default dispatcher so wrapping for context propagation
+		// preserves the standard "OkHttp Dispatcher-N" thread names.
+		private static ExecutorService defaultDispatcherExecutor() {
+			AtomicInteger counter = new AtomicInteger(1);
+			ThreadFactory threadFactory = runnable -> {
+				Thread thread = new Thread(runnable, "OkHttp Dispatcher-" + counter.getAndIncrement());
+				thread.setDaemon(false);
+				return thread;
+			};
+			return new ThreadPoolExecutor(0, Integer.MAX_VALUE, 60L, TimeUnit.SECONDS, new SynchronousQueue<>(),
+					threadFactory);
 		}
 
 	}
