+++ spring-ai-commons/pom.xml
@@ -83,13 +83,6 @@
 			<version>${jtokkit.version}</version>
 		</dependency>
 
-		<dependency>
-			<groupId>com.squareup.okhttp3</groupId>
-			<artifactId>okhttp</artifactId>
-			<version>${okhttp3.version}</version>
-			<optional>true</optional>
-		</dependency>
-
 		<!-- test dependencies -->
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
